requirements.txt
# Core
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0

# Config
hydra-core>=1.3.0
omegaconf>=2.3.0

# Entity / constituency parsing
spacy>=3.5.0
benepar>=0.2.0
stanza>=1.5.0

# PEFT / fine-tuning (for ft_attn_flow_test.py)
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0

# Data & utils
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
